move validations to gitops folder
