+++ src/providers/gitops/api.js
@@ -1,20 +1,13 @@
-const fs = require('fs')
 const config = require('../../config')
-const { deployHooksSchema } = require('../../validations/deploy-hooks.validation')
-
-const _getValidatedDeploymentsConfig =  async () => {
-    const deploymentsConfigFile = await fs.promises.readFile(config.DEPLOYMENTS_CONFIG_PATH, 'utf8');
-    const deploymentsConfigJson = JSON.parse(deploymentsConfigFile)
-    return deployHooksSchema.validateAsync(deploymentsConfigJson)
-}
+const { getValidatedDeploymentsConfig } = require('./validations')
 
 const findAllDeploymentApps = async () => {
-    const deployments = await _getValidatedDeploymentsConfig()
+    const deployments = await getValidatedDeploymentsConfig(config.DEPLOYMENTS_CONFIG_PATH)
     return deployments.apps
 }
 
 const findOneDeploymentApp = async (appName) => {
-    const deployments = await _getValidatedDeploymentsConfig()
+    const deployments = await getValidatedDeploymentsConfig(config.DEPLOYMENTS_CONFIG_PATH)
     return deployments.apps.find(app => app.name === appName)
 }
